notifications/requirements.txt
psycopg2-binary
firebase-admin